1. 토큰 재발급 추가 2. 회원 가입시 문구 수정 3. 회원 탈퇴 추가 4. user 모듈 patch 오류 수정 5. user 모듈 삭제시 예외 추가

+++ src/modules/users/users.service.ts
@@ -120,6 +120,9 @@ export class UsersService {
      */
     async softDeleteUser(uid: number):Promise<UpdateResult>{
         try{
+            const user = await this.usersRepository.findOneBy({ uid });
+            if (!user) throw new NotFoundException("존재하지 않는 회원의 정보를 삭제할 수 없습니다.");
+
             return await this.usersRepository.softDelete({uid});
         }catch (error){
             this.logger.error(error);
@@ -135,6 +138,9 @@ export class UsersService {
      */
     async hardDeleteUser(uid: number): Promise<DeleteResult>{
         try{
+            const user = await this.usersRepository.findOneBy({ uid });
+            if (!user) throw new NotFoundException("존재하지 않는 회원의 정보를 삭제할 수 없습니다.");
+
             return await this.usersRepository.delete({uid});
         }catch (error){
             this.logger.error(error);
